+++ Coimbra.Editor.Linting/AssemblyDefinitionRuleBase.cs
@@ -1,5 +1,4 @@
-﻿using Coimbra.Editor;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -11,7 +10,7 @@ using UnityEngine;
 using UnityEngine.Serialization;
 using Object = UnityEngine.Object;
 
-namespace Coimbra.Linting.Editor
+namespace Coimbra.Editor.Linting
 {
     /// <summary>
     /// Base class to define a new rule for any <see cref="AssemblyDefinitionAsset"/>.